Evaluation Metrics

Evaluation Metrics

TabularReinforcementLearning.AllRewards — Type.

struct AllRewards <: AbstractEvaluationMetrics
	rewards::Array{Float64, 1}

Records all rewards.

source

TabularReinforcementLearning.AllRewards — Method.

AllRewards()

Initializes with empty array.

source

TabularReinforcementLearning.EvaluationPerEpisode — Type.

EvaluationPerEpisode <: AbstractEvaluationMetrics
	values::Array{Float64, 1}
	metric::SimpleEvaluationMetric

Stores the value of the simple metric for each episode in values.

source

TabularReinforcementLearning.EvaluationPerEpisode — Type.

EvaluationPerEpisode(metric = MeanReward())

Initializes with empty values array and simple metric (default MeanReward). Other options are TimeSteps (to measure the lengths of episodes) or TotalReward.

source

TabularReinforcementLearning.EvaluationPerT — Type.

EvaluationPerT <: AbstractEvaluationMetrics
	T::Int64
	counter::Int64
	values::Array{Float64, 1}
	metric::SimpleEvaluationMetric

Stores the value of the simple metric after every T steps in values.

source

TabularReinforcementLearning.EvaluationPerT — Type.

EvaluationPerT(T, metric = MeanReward())

Initializes with T, counter = 0, empty values array and simple metric (default MeanReward). Another option is TotalReward.

source

TabularReinforcementLearning.MeanReward — Type.

mutable struct MeanReward <: TabularReinforcementLearning.SimpleEvaluationMetric
	meanreward::Float64
	counter::Int64

Computes iteratively the mean reward.

source

TabularReinforcementLearning.MeanReward — Method.

MeanReward()

Initializes counter and meanreward to 0.

source

TabularReinforcementLearning.RecordAll — Type.

struct RecordAll <: AbstractEvaluationMetrics
	r::Array{Float64, 1}
	a::Array{Int64, 1}
	s::Array{Int64, 1}
	isterminal::Array{Bool, 1}

Records everything.

source

TabularReinforcementLearning.RecordAll — Method.

RecordAll()

Initializes with empty arrays.

source

TabularReinforcementLearning.TimeSteps — Type.

mutable struct TimeSteps <: SimpleEvaluationMetric
	counter::Int64

Counts the number of timesteps the simulation is running.

source

TabularReinforcementLearning.TimeSteps — Method.

TimeSteps()

Initializes counter to 0.

source

TabularReinforcementLearning.TotalReward — Type.

mutable struct TotalReward <: TabularReinforcementLearning.SimpleEvaluationMetric
	reward::Float64

Accumulates all rewards.

source

TabularReinforcementLearning.TotalReward — Method.

TotalReward()

Initializes reward to 0.

source